node-main/deps/v8/test/inspector/regress/regress-crbug-1290861-expected.txt

Ensure that catch prediction is correct for [[Reject]] handlers.

Running test: test
Uncaught exception in rejectHandler
